\name{stSemiPar-package}
\alias{stSemiPar-package}
\alias{stSemiPar}
\docType{package}
\title{
\packageTitle{stSemiPar}
}
\description{
\packageDescription{stSemiPar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stSemiPar}
\packageIndices{stSemiPar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stSemiPar}

Maintainer: \packageMaintainer{stSemiPar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
